feat: new association between card & player

+++ src/main/java/com/site/card/site/dto/CardDTO.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 
 public class CardDTO {
     @Getter @Setter
-    private int id;
+    private long id;
     @Getter @Setter
     private String name;
     @Getter @Setter
@@ -23,7 +23,7 @@ public class CardDTO {
     @Getter @Setter
     private String url;
 
-    public CardDTO(int id, String name, String description, String affinity, String family, float energy, float hp, float price) {
+    public CardDTO(long id, String name, String description, String affinity, String family, float energy, float hp, float price) {
         this.id = id;
         this.name = name;
         this.description = description;
